Session Handling
Mike & Pat

If you want to make sure a user is logged in this snippet of PHP has to
and I mean MUST come before the HTML tag.  It checks to see if there is
a username in the session. if there is not, it sends them back to the
login page to login. This will work with our current time out expiration
and does not need any cookies.